requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2
django-crispy-forms==1.14.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing-extensions==4.7.1
qrcode==7.3.1
fcm-django==1.0.12
pytz==2023.4

django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.6
sqlparse==0.4.4
typing-extensions==4.7.1
APScheduler==3.10.4





